added check before removal
git-svn-id: https://svn.resiprocate.org/rep/resiprocate@3011 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ main/sip/use-autotools.sh
@@ -111,7 +111,7 @@ find . -name ${ATCF} -exec cp /dev/null {} \;
 # Remove conflicting files -- and add to ignore list.
 for F in ${AUTOTOOLS_CONFLICTS} ${AUTOTOOLS_USED}; do
     for E in echo "" ; do
-        ${E} rm ${F}
+        [ -e "${F}" ] && ${E} rm ${F}
     done
     Fd=$(dirname ${F})
     Fb=$(basename ${F})
